Changed exponent representation from Int to custom WExp type, which for now simply wraps an Int, as a preliminary step toward moving the sign bit to the exponent.

+++ Tests/BigMathTests/WideFloatTests/WideFloat_Buffer_UnitTests.swift
@@ -39,7 +39,7 @@ class WideFloat_Buffer_UnitTests: XCTestCase
         {
             var buff = FloatingPointBuffer(wideFloatUIntBuffer: $0[...])
             buff.rightShiftForAddOrSubtract(by: shift)
-            exponent = buff.exponent
+            exponent = buff.exponent.intValue
         }
         
         return (x.dropLast(), exponent)
